ControlEditBug AllowEdit Bugfix, Misc
Corrected a bug with the private function UpdateForAllowEditing; If
Allow editing was false, it would allow editing, while if it was true,
it would not allow for editing (the exact opposite of what we want)

Also added initializing default values to the constructor, for
completion's sake (and safety).

Added a few files to the gitignore, since we don't need to commit
executables, pdbs, or in general, th entire *\Debug and *\obj folders,
since they are compiled by Visual Studio.